Extract WatirOptionsFactory class from Browser#initialize
